2026/day-36/python-food-app/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.1
WTForms==3.1.1
Werkzeug==3.0.1
SQLAlchemy==2.0.48
email-validator==2.1.0
PyMySQL==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
gunicorn
